Fix bug where publishing first message to topic took 1 second
There is a bug in librdkafka where producing the very first message into
a topic always takes around 1 second.

This happens due to bug that was first raised in this GitHub issue:
https://github.com/confluentinc/librdkafka/issues/3202

The maintainer proposed a fix for this problem and said it was going to
be fixed starting librdkafka 1.7.0, but it seems the fix was actually
never merged since the problem still existed. I think that the
maintainer mixed something up there because they actually closed the
issue, perhaps they forgot to merge the fix or something similar.

This commit makes rdkafka use our fork of librdkafka 2.3.0. This fork is
almost identical to the regular 2.3.0 but it has the same fix applied
that the maintainer proposed.

You can find the commit here:
https://github.com/salemove/librdkafka/commit/71eea91d3d52f9738594a75aa6f098545ee0055a

XT-59